+++ src/main/java/ru/yandex/practicum/javafilmorate/storage/interfaceStorage/UserStorage.java
@@ -13,14 +13,14 @@ public interface UserStorage {
 
     void updateUser(User user);
 
-    Optional<User> getUser(int userId);
+    Optional<User> getUser(Long userId);
 
-    List<User> getUsersFriends(Integer id);
+    List<User> getUsersFriends(Long id);
 
-    List<User> getCommonFriends(Integer userId, Integer friendId);
+    List<User> getCommonFriends(Long userId, Long friendId);
 
-    void deleteFriend(Integer userId, Integer friendId);
+    void deleteFriend(Long userId, Long friendId);
 
-    void addFriend(Integer userId, Integer friendId);
+    void addFriend(Long userId, Long friendId);
 
 }
